stub-generator: Skip setting `'kind'` on `Namespace_` nodes (#36871)
We had been setting the `'kind'` to try to use braced namespaces unless
there's only one. But it turns out the printer has its own logic, it
ignores the kind and figures out on its own whether it needs to used
braced namespaces or not. So there's no point in us bothering to set the
`'kind'`, and it may even be confusing when we set it to braced but the
printer uses semicolons anyway.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/8658881774